refactor(bin-designer): split useCutoutInteraction into focused sub-hooks
Decompose the 1213-LOC useCutoutInteraction.ts into 9 sibling files by
concern. The parent hook is now a thin orchestrator (429 LOC) that holds
the shared state machine and composes domain-specific sub-hooks.

Sibling files (all under 500 LOC):

- cutoutInteractionTypes.ts (119) — InteractionMode discriminated union,
  ResizeHandle, PreviewMap, UseCutoutInteractionOptions, default size
  constants.
- cutoutHelpers.ts (96) — cloneCutoutsWithGroups, clampedOffset,
  addClonedCutouts, createDefaultCutout. Pure helpers, no React.
- useCutoutClipboard.ts (105) — clipboard state + copy/paste/duplicate
  with mask-aware placement.
- useCutoutContextMenu.ts (28) — right-click menu state.
- useCutoutUndoToasts.ts (46) — toast-wrapped undo/redo.
- useCutoutKeyboardShortcuts.ts (134) — keydown listener effect.
- useCutoutRecovery.ts (93) — pointercancel/blur/visibility resets +
  ruler-measurement cleanup on mode change.
- useCutoutTransformStarters.ts (175) — startDrag/startResize/
  startRotation/startGroupRotation/startGroupScale with locked-cutout
  guards and group-aware drag selection.
- useCutoutPathTool.ts (238) — path drawing + vertex editing handlers
  (commitPath, handlePathBackgroundDown, vertex pointer/handle/background
  down, enterVertexEditing).
- useCutoutPointerHandlers.ts (376) — handlePointerMove dispatcher and
  handlePointerUp with the three commit functions
  (commitPendingPlace/commitTransformPreview/commitDrawing).

Public API unchanged: the hook's return shape is identical, so
CutoutWorkspace.tsx and other consumers are not modified. Type re-exports
(`ResizeHandle`, `InteractionMode`, `PreviewMap`) preserved at the parent
hook for module-path compatibility.

Also clean up an unused `react-hooks/set-state-in-effect` disable that
the rule no longer flags after the recovery effect's restructuring,
and replace deprecated `React.MutableRefObject` with `React.RefObject`
on the new sub-hook signatures.

useCutoutInteraction.ts: 1213 LOC -> 429 LOC. All 10,278 tests pass.
